Separate modules into main, demo, testbench and integration-tests
