+++ vaadin-radio-button-flow/src/test/java/com/vaadin/flow/component/radiobutton/tests/RadioButtonGroupTest.java
@@ -13,11 +13,12 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow.component.radiobutton;
+package com.vaadin.flow.component.radiobutton.tests;
 
 import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.HasValue;
 import com.vaadin.flow.component.HasValue.ValueChangeEvent;
+import com.vaadin.flow.component.radiobutton.RadioButtonGroup;
 import org.junit.Assert;
 import org.junit.Test;
